Feat: add CRUD features
In this app, any user can create a new kitten, edit it, delete it and
view a list of all the created but not deleted kittens. There are links
to edit and delete a kitten in the index and show pages. In the edit
page, there is a link to destroy the kitten.

Also there are flash messages that display a message when the user
fails to create or update a kitten, or when it achieves to create,
update or destroy a kitten.